+++ Tests/Reqnroll.RuntimeTests/TestRunnerManagerTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Globalization;
+using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
 using FluentAssertions;
@@ -27,7 +28,7 @@ public class TestRunnerManagerTests : IAsyncLifetime
     [Fact]
     public void CreateTestRunner_should_be_able_to_create_a_TestRunner()
     {
-        var testRunner = _testRunnerManager.CreateTestRunner();
+        var testRunner = _testRunnerManager.GetOrCreateTestRunner();
 
         testRunner.Should().NotBeNull();
         testRunner.Should().BeOfType<TestRunner>();
@@ -71,6 +72,95 @@ public class TestRunnerManagerTests : IAsyncLifetime
         TestRunnerManager.ReleaseTestRunner(testRunner3);
     }
 
+    [Fact]
+    public void Should_use_test_runner_that_was_active_last_time_on_the_same_thread_if_possible()
+    {
+        var ourFeatureInfo = new FeatureInfo(new CultureInfo("en-US"), null, "F1", null);
+        var otherFeatureInfo = new FeatureInfo(new CultureInfo("en-US"), null, "F2", null);
+
+        var otherRunners1 = Enumerable.Range(0, 5).Select(_ => TestRunnerManager.GetTestRunnerForAssembly(featureHint: otherFeatureInfo)).ToList();
+        var otherRunners2 = Enumerable.Range(0, 5).Select(_ => TestRunnerManager.GetTestRunnerForAssembly(featureHint: otherFeatureInfo)).ToList();
+        var runnerOnDifferentFeatureSameThread = TestRunnerManager.GetTestRunnerForAssembly(featureHint: otherFeatureInfo);
+        var runnerOnSameFeatureDifferentThread = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+        var ourRunner = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+
+        void RunOnOtherThreadAndWait(Action action)
+        {
+            Task.Run(action).Wait(500);
+        }
+
+        // release otherRunners1 on a different thread
+        RunOnOtherThreadAndWait(() => otherRunners1.ForEach(TestRunnerManager.ReleaseTestRunner));
+        
+        // release the selected runners
+        TestRunnerManager.ReleaseTestRunner(ourRunner);
+        TestRunnerManager.ReleaseTestRunner(runnerOnDifferentFeatureSameThread);
+        RunOnOtherThreadAndWait(() => TestRunnerManager.ReleaseTestRunner(runnerOnSameFeatureDifferentThread));
+
+        // release otherRunners2 on a different thread
+        RunOnOtherThreadAndWait(() => otherRunners2.ForEach(TestRunnerManager.ReleaseTestRunner));
+
+        // Priority:
+        // 1. Same feature, same thread
+        // 2. Same feature, different thread
+        // 3. Other feature, same thread
+        // 4. Other feature, other thread
+
+        // Priority 1: from the same feature & same thread, we should get our test runner first
+        var ourRunnerAgain = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+        ourRunnerAgain.TestWorkerId.Should().Be(ourRunner.TestWorkerId);
+
+        // Priority 2: from the same feature & different thread
+        var runnerOnSameFeatureDifferentThreadAgain = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+        runnerOnSameFeatureDifferentThreadAgain.TestWorkerId.Should().Be(runnerOnSameFeatureDifferentThread.TestWorkerId);
+
+        // Priority 3: from the different feature & same thread
+        var runnerOnDifferentFeatureSameThreadAgain = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+        runnerOnDifferentFeatureSameThreadAgain.TestWorkerId.Should().Be(runnerOnDifferentFeatureSameThread.TestWorkerId);
+
+        // Priority 4: if there is no from the same thread, just provide one of the others
+        var otherRunnerWithoutHint = TestRunnerManager.GetTestRunnerForAssembly(featureHint: ourFeatureInfo);
+        otherRunnerWithoutHint.TestWorkerId.Should().NotBe(ourRunner.TestWorkerId);
+
+        TestRunnerManager.ReleaseTestRunner(ourRunnerAgain);
+        TestRunnerManager.ReleaseTestRunner(runnerOnSameFeatureDifferentThreadAgain);
+        TestRunnerManager.ReleaseTestRunner(runnerOnDifferentFeatureSameThreadAgain);
+        TestRunnerManager.ReleaseTestRunner(otherRunnerWithoutHint);
+    }
+
+    [Fact]
+    public async Task Should_fire_remaining_AfterFeature_hooks_of_test_threads()
+    {
+        _testRunnerManager.Initialize(_anAssembly);
+
+        var testRunnerWithFeatureContext = _testRunnerManager.GetTestRunner();
+        var testRunnerWithoutFeatureContext = _testRunnerManager.GetTestRunner();
+
+        FeatureHookTracker.AfterFeatureCalled = false;
+        await testRunnerWithFeatureContext.OnFeatureStartAsync(new FeatureInfo(new CultureInfo("en-US", false), string.Empty, "F", null));
+        var disposableClass1 = new DisposableClass();
+        testRunnerWithFeatureContext.FeatureContext.FeatureContainer.RegisterInstanceAs(disposableClass1, dispose: true);
+
+        TestRunnerManager.ReleaseTestRunner(testRunnerWithFeatureContext);
+        TestRunnerManager.ReleaseTestRunner(testRunnerWithoutFeatureContext);
+
+        await TestRunnerManager.OnTestRunEndAsync(_anAssembly);
+        FeatureHookTracker.AfterFeatureCalled.Should().BeTrue();
+        disposableClass1.IsDisposed.Should().BeTrue();
+    }
+
+    [Binding]
+    static class FeatureHookTracker
+    {
+        public static bool AfterFeatureCalled = false;
+
+        [AfterFeature]
+        public static void AfterFeature()
+        {
+            AfterFeatureCalled = true;
+        }
+    }
+
     class DisposableClass : IDisposable
     {
         public bool IsDisposed { get; private set; }
@@ -106,7 +196,7 @@ public class TestRunnerManagerTests : IAsyncLifetime
     {
         _testRunnerManager.IsTestRunInitialized.Should().BeFalse("binding registry should not be initialized initially");
 
-        _testRunnerManager.CreateTestRunner();
+        _testRunnerManager.GetOrCreateTestRunner();
 
         _testRunnerManager.IsTestRunInitialized.Should().BeTrue("binding registry be initialized");
     }