+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # [vNext]
 
 ## Improvements:
+
 * Upgrade to Gherkin v30 from v29 (see [Gherkin changelog](https://github.com/cucumber/gherkin/blob/main/CHANGELOG.md)) (#305)
+* Support scenario-level (method-level) parallel execution (#119, #277)
 
 ## Bug fixes: